proto/requirements.txt
hidapi>=0.14.0
bleak>=0.21.0
pyusb>=1.3.0
libusb>=1.0.27
